Add method to check if we have mapped a port. Check that port is mapped before deleting the mapping in Close Lobby

+++ game/src/main/java/org/dragonskulle/game/lobby/Lobby.java
@@ -213,7 +213,9 @@ public class Lobby extends Component implements IFrameUpdate {
                                     LobbyDiscovery.closeLocalLobby();
                                     mHostUi.setEnabled(true);
                                     mHostingUi.setEnabled(false);
-                                    UPnP.deletePortMapping(PORT, "TCP");
+                                    if (UPnP.checkMappingExists(PORT, "TCP")) {
+                                        UPnP.deletePortMapping(PORT, "TCP");
+                                    }
                                 }));
 
         UIManager.getInstance()
